+++ threedigrid_builder/tests/test_embedded.py
@@ -180,6 +180,9 @@ def test_embed_linear_objects_multiple(grid2d):
         ([(5, 5), (5, 10), (7, 10), (7, 15), (10, 15), (10, 17), (15, 17)], None, [2]),
         ([(5.6, 7), (10, 10.3), (15, 10.3)], [5.25], []),  # 0 to 3, 2 is below thresh
         ([(5.6, 7), (10, 10.3), (14.4, 7)], [5.5], []),  # 0 to 1, 2&3 are below thresh
+        ([(0, 5), (18, 5)], [10], []),  # begins at model edge
+        ([(0, 5), (18, 5), (18, 12)], [10, 23], [1]),  # begins at model edge, with node
+        ([(-1, 5), (18, 5)], [11], []),  # begins outside of model
     ],
 )
 def test_embed_linear_object(grid2d, geometry, lines_s1d, embedded_in, reverse):